# book-network-ui/src/app/services/models/register-user-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface RegisterUserDto {
  dateOfBirth?: string;
  email?: string;
  firstName: string;
  lastName?: string;
  password: string;
}